WIP - Refactor for FormAliasBaseComponent
