+++ cypress/parallel/weights/functional-parallel-weights.json
@@ -1,1 +1,1 @@
-{"cypress/e2e/functional/opal/PO-117_accounEnquirytDetails.feature":{"time":11978,"weight":2},"cypress/e2e/functional/opal/PO-136_PO-137_accountHistory.feature":{"time":6911,"weight":1},"cypress/e2e/functional/opal/PO-251_unauthorisedUser.feature":{"time":1674,"weight":0},"cypress/e2e/functional/opal/account_details/PO-366_updateAccountDetails.feature":{"time":13356,"weight":2},"cypress/e2e/functional/opal/account_details/PO-367_accountCreation_parentOrGaurdian.feature":{"time":13589,"weight":2},"cypress/e2e/functional/opal/manualAccountCreation/PO-280_employerDetails.feature":{"time":50074,"weight":10},"cypress/e2e/functional/opal/manualAccountCreation/PO-343_manualAccountCreation.feature":{"time":1896,"weight":0},"cypress/e2e/functional/opal/manualAccountCreation/PO-346_Business_Unit_And_Defendant_type.feature":{"time":19489,"weight":4},"cypress/e2e/functional/opal/manualAccountCreation/PO-347_createAccountUserAccessOneBU.feature":{"time":6892,"weight":1},"cypress/e2e/functional/opal/manualAccountCreation/PO-362_companyTaskList.feature":{"time":21288,"weight":4},"cypress/e2e/functional/opal/manualAccountCreation/PO-364_parentOrGuardianToPayPage.feature":{"time":63327,"weight":13},"cypress/e2e/functional/opal/manualAccountCreation/PO-368_employerDetails_parentAndGuardian.feature":{"time":59122,"weight":12},"cypress/e2e/functional/opal/manualAccountCreation/PO-377_createAccountUserAccessMultipleBU.feature":{"time":29546,"weight":6},"cypress/e2e/functional/opal/manualAccountCreation/adultOrYouthOnly/PO-360_personalDetails.feature":{"time":139268,"weight":29},"cypress/e2e/functional/opal/manualAccountCreation/adultOrYouthOnly/PO-433_addContactDetailsButton.feature":{"time":119131,"weight":25},"cypress/e2e/functional/opal/manualAccountCreation/adultOrYouthWithParentOrGuardian/PO-369_personalDetailsParentOrGuardian.feature":{"time":202782,"weight":43},"cypress/e2e/functional/opal/manualAccountCreation/contactDetails/PO-370_contactDetails_parentOrGuardian.feature":{"time":46173,"weight":9},"cypress/e2e/functional/opal/manualAccountCreation/contactDetails/PO-371_contactDetails_company.feature":{"time":46023,"weight":9},"cypress/e2e/functional/opal/manualAccountCreation/contactDetails/PO-419_contactDetails_adultOrYouth.feature":{"time":43266,"weight":9}}
+{"cypress/e2e/functional/opal/PO-117_accounEnquirytDetails.feature":{"time":12015,"weight":2},"cypress/e2e/functional/opal/PO-136_PO-137_accountHistory.feature":{"time":7213,"weight":1},"cypress/e2e/functional/opal/PO-251_unauthorisedUser.feature":{"time":1841,"weight":0},"cypress/e2e/functional/opal/account_details/PO-366_updateAccountDetails.feature":{"time":13278,"weight":2},"cypress/e2e/functional/opal/account_details/PO-367_accountCreation_parentOrGaurdian.feature":{"time":13552,"weight":2},"cypress/e2e/functional/opal/manualAccountCreation/PO-280_employerDetails.feature":{"time":50293,"weight":10},"cypress/e2e/functional/opal/manualAccountCreation/PO-343_manualAccountCreation.feature":{"time":1972,"weight":0},"cypress/e2e/functional/opal/manualAccountCreation/PO-346_Business_Unit_And_Defendant_type.feature":{"time":19059,"weight":4},"cypress/e2e/functional/opal/manualAccountCreation/PO-347_createAccountUserAccessOneBU.feature":{"time":6892,"weight":1},"cypress/e2e/functional/opal/manualAccountCreation/PO-362_companyTaskList.feature":{"time":21115,"weight":4},"cypress/e2e/functional/opal/manualAccountCreation/PO-364_parentOrGuardianToPayPage.feature":{"time":64822,"weight":13},"cypress/e2e/functional/opal/manualAccountCreation/PO-368_employerDetails_parentAndGuardian.feature":{"time":59594,"weight":12},"cypress/e2e/functional/opal/manualAccountCreation/PO-377_createAccountUserAccessMultipleBU.feature":{"time":29628,"weight":6},"cypress/e2e/functional/opal/manualAccountCreation/adultOrYouthOnly/PO-360_personalDetails.feature":{"time":140214,"weight":29},"cypress/e2e/functional/opal/manualAccountCreation/adultOrYouthOnly/PO-433_addContactDetailsButton.feature":{"time":120639,"weight":25},"cypress/e2e/functional/opal/manualAccountCreation/adultOrYouthWithParentOrGuardian/PO-369_personalDetailsParentOrGuardian.feature":{"time":204876,"weight":43},"cypress/e2e/functional/opal/manualAccountCreation/contactDetails/PO-370_contactDetails_parentOrGuardian.feature":{"time":46098,"weight":9},"cypress/e2e/functional/opal/manualAccountCreation/contactDetails/PO-371_contactDetails_company.feature":{"time":45217,"weight":9},"cypress/e2e/functional/opal/manualAccountCreation/contactDetails/PO-419_contactDetails_adultOrYouth.feature":{"time":42826,"weight":9}}
