+++ cypress/parallel/weights/smoke-parallel-weights.json
@@ -1,1 +1,1 @@
-{"cypress/e2e/smoke/opal/PO-151_signOutTest.feature":{"time":2869,"weight":3},"cypress/e2e/smoke/opal/PO-224_accessibilityTests.feature":{"time":22181,"weight":24},"cypress/e2e/smoke/opal/signInTest.feature":{"time":2520,"weight":2}}
+{"cypress/e2e/smoke/opal/PO-151_signOutTest.feature":{"time":2460,"weight":2},"cypress/e2e/smoke/opal/PO-224_accessibilityTests.feature":{"time":22780,"weight":25},"cypress/e2e/smoke/opal/signInTest.feature":{"time":2067,"weight":2}}
