+++ cypress/support/projectConfig/personal_details_page.ts
@@ -24,13 +24,13 @@ export default class personalDetails {
     cy.get('#addressLine1').find('input').clear().type(addLine3);
   }
   static enterFirstNamesInAlias(firstNameAlias: string) {
-    cy.get('#addNameAlias-conditional > fieldset >legend')
+    cy.get('#addAlias-conditional > fieldset >legend')
       .siblings()
-      .get('#addNameAlias-conditional > fieldset >app-govuk-text-input > div > input')
+      .get('#addAlias-conditional > fieldset >app-govuk-text-input > div > input')
       .should('be.visible')
       .type(firstNameAlias);
   }
   static enterLastNamesInAlias(lastNameAlias: string) {
-    cy.get('#addNameAlias-conditional > fieldset >app-govuk-text-input >div>input').type(lastNameAlias);
+    cy.get('#addAlias-conditional > fieldset >app-govuk-text-input >div>input').type(lastNameAlias);
   }
 }
