+++ src/app/constants/manual-account-creation-company-details-state.ts
@@ -1,7 +1,7 @@
 export const MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_STATE = {
   companyName: null,
-  addCompanyAlias: null,
-  companyAliases: [],
+  addAlias: null,
+  aliases: [],
   addressLine1: null,
   addressLine2: null,
   addressLine3: null,
