+++ src/app/constants/manual-account-creation-personal-details-state.ts
@@ -2,8 +2,8 @@ export const MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_STATE = {
   title: null,
   firstNames: null,
   lastName: null,
-  addNameAlias: null,
-  nameAliases: [],
+  addAlias: null,
+  aliases: [],
   dateOfBirth: null,
   nationalInsuranceNumber: null,
   addressLine1: null,
