+++ src/app/interfaces/manual-account-creation-personal-details-state.interface.ts
@@ -4,8 +4,8 @@ export interface IManualAccountCreationPersonalDetailsState {
   title: string | null;
   firstNames: string | null;
   lastName: string | null;
-  addNameAlias: boolean | null;
-  nameAliases: IManualAccountCreationPersonalDetailsAliasState[];
+  addAlias: boolean | null;
+  aliases: IManualAccountCreationPersonalDetailsAliasState[];
   dateOfBirth: string | null;
   nationalInsuranceNumber: string | null;
   addressLine1: string | null;
