+++ src/app/pages/manual-account-creation/personal-details/personal-details.component.spec.ts
@@ -22,8 +22,8 @@ describe('PersonalDetailsComponent', () => {
       title: 'Mr',
       firstNames: null,
       lastName: null,
-      addNameAlias: false,
-      nameAliases: [],
+      addAlias: false,
+      aliases: [],
       dateOfBirth: null,
       nationalInsuranceNumber: null,
       addressLine1: null,
